cleanup, move things around
