+++ docs/source/api/apollo-gateway.mdx
@@ -100,7 +100,7 @@ const gateway = new ApolloGateway({
 
 ###### `supergraphSdl`
 
-`string | SupergraphSdlHook | SupergraphSdlManager`
+`string | SupergraphSdlHook | SupergraphManager`
 </td>
 <td>
 
@@ -111,7 +111,7 @@ When `supergraphSdl` is a `SupergraphSdlHook`: An `async` function that returns 
 2. `healthCheck`: A function which issues a health check against the subgraphs
 3. `getDataSource`: A function which gets a data source for a particular subgraph from the gateway
 
-When `supergraphSdl` is a `SupergraphSdlManager`: An object containing an `initialize` property. `initialize` is an `async` function of the `SupergraphSdlHook` type mentioned directly above.
+When `supergraphSdl` is a `SupergraphManager`: An object containing an `initialize` property. `initialize` is an `async` function of the `SupergraphSdlHook` type mentioned directly above.
 
 **If you are using managed federation,** do not provide this field.
