--- gateway-js/src/IntrospectAndCompose/__tests__/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../../tsconfig.test",
-  "include": ["**/*"],
-  "references": [
-    { "path": "../../../" },
-    { "path": "../../../../federation-integration-testsuite-js" },
-  ]
-}
