+++ gateway-js/src/supergraphManagers/IntrospectAndCompose/__tests__/IntrospectAndCompose.test.ts
@@ -4,7 +4,7 @@ import {
   fixturesWithUpdate,
 } from 'apollo-federation-integration-testsuite';
 import { nockBeforeEach, nockAfterEach } from '../../__tests__/nockAssertions';
-import { RemoteGraphQLDataSource, ServiceEndpointDefinition } from '../..';
+import { RemoteGraphQLDataSource, ServiceEndpointDefinition } from '../../..';
 import { IntrospectAndCompose } from '..';
 import { mockAllServicesSdlQuerySuccess } from '../../__tests__/integration/nockMocks';
 import { getTestingSupergraphSdl, wait } from '../../__tests__/execution-utils';
