+++ gateway-js/src/supergraphManagers/UplinkFetcher/__tests__/loadSupergraphSdlFromStorage.test.ts
@@ -2,7 +2,7 @@ import {
   loadSupergraphSdlFromStorage,
   loadSupergraphSdlFromUplinks
 } from '../loadSupergraphSdlFromStorage';
-import { getDefaultFetcher } from '../..';
+import { getDefaultFetcher } from '../../..';
 import {
   graphRef,
   apiKey,
@@ -14,9 +14,9 @@ import {
   mockSupergraphSdlRequestSuccess,
   mockSupergraphSdlRequestIfAfterUnchanged,
   mockSupergraphSdlRequestIfAfter
-} from '../../__tests__/integration/nockMocks';
-import { getTestingSupergraphSdl } from "../../__tests__/execution-utils";
-import { nockAfterEach, nockBeforeEach } from '../../__tests__/nockAssertions';
+} from '../../../__tests__/integration/nockMocks';
+import { getTestingSupergraphSdl } from "../../../__tests__/execution-utils";
+import { nockAfterEach, nockBeforeEach } from '../../../__tests__/nockAssertions';
 
 describe('loadSupergraphSdlFromStorage', () => {
   beforeEach(nockBeforeEach);
