+++ gateway-js/src/supergraphManagers/index.ts
@@ -1,0 +1,4 @@
+export { LocalCompose } from './LocalCompose';
+export { LegacyFetcher } from './LegacyFetcher';
+export { IntrospectAndCompose } from './IntrospectAndCompose';
+export { UplinkFetcher } from './UplinkFetcher';
